refactor: add ecash and onchain subcommands
