+++ gateway/cli/src/general_commands.rs
@@ -1,15 +1,11 @@
 use anyhow::bail;
-use bitcoin::address::NetworkUnchecked;
-use bitcoin::Address;
 use clap::Subcommand;
 use fedimint_core::config::FederationId;
-use fedimint_core::{fedimint_build_code_version_env, Amount, BitcoinAmountOrAll};
-use fedimint_mint_client::OOBNotes;
+use fedimint_core::fedimint_build_code_version_env;
 use ln_gateway::rpc::rpc_client::GatewayRpcClient;
 use ln_gateway::rpc::{
-    BackupPayload, BalancePayload, ConfigPayload, ConnectFedPayload, DepositAddressPayload,
-    FederationRoutingFees, LeaveFedPayload, ReceiveEcashPayload, SetConfigurationPayload,
-    SpendEcashPayload, WithdrawPayload,
+    BalancePayload, ConfigPayload, ConnectFedPayload, FederationRoutingFees, LeaveFedPayload,
+    SetConfigurationPayload,
 };
 
 use crate::print_response;
@@ -60,23 +56,6 @@ pub enum GeneralCommands {
     },
     /// Get the total on-chain, lightning, and eCash balances of the gateway.
     GetBalances,
-    /// Generate a new peg-in address to a federation that the gateway can claim
-    /// e-cash for later.
-    Address {
-        #[clap(long)]
-        federation_id: FederationId,
-    },
-    /// Claim funds from a gateway federation to an on-chain address.
-    Withdraw {
-        #[clap(long)]
-        federation_id: FederationId,
-        /// The amount to withdraw
-        #[clap(long)]
-        amount: BitcoinAmountOrAll,
-        /// The address to send the funds to
-        #[clap(long)]
-        address: Address<NetworkUnchecked>,
-    },
     /// Register the gateway with a federation.
     ConnectFed {
         /// Invite code to connect to the federation
@@ -91,11 +70,6 @@ pub enum GeneralCommands {
     },
     /// Leave a federation.
     LeaveFed {
-        #[clap(long)]
-        federation_id: FederationId,
-    },
-    /// Make a backup of snapshot of all e-cash.
-    Backup {
         #[clap(long)]
         federation_id: FederationId,
     },
@@ -124,25 +98,6 @@ pub enum GeneralCommands {
     },
     /// Safely stop the gateway
     Stop,
-    /// Spend e-cash
-    SpendEcash {
-        #[clap(long)]
-        federation_id: FederationId,
-        amount: Amount,
-        #[clap(long)]
-        allow_overpay: bool,
-        #[clap(long, default_value_t = 60 * 60 * 24 * 7)]
-        timeout: u64,
-        #[clap(long)]
-        include_invite: bool,
-    },
-    /// Receive e-cash
-    ReceiveEcash {
-        #[clap(long)]
-        notes: OOBNotes,
-        #[arg(long = "no-wait", action = clap::ArgAction::SetFalse)]
-        wait: bool,
-    },
 }
 
 impl GeneralCommands {
@@ -186,28 +141,6 @@ impl GeneralCommands {
                 let response = create_client().get_balances().await?;
                 print_response(response);
             }
-            Self::Address { federation_id } => {
-                let response = create_client()
-                    .get_deposit_address(DepositAddressPayload { federation_id })
-                    .await?;
-
-                print_response(response);
-            }
-            Self::Withdraw {
-                federation_id,
-                amount,
-                address,
-            } => {
-                let response = create_client()
-                    .withdraw(WithdrawPayload {
-                        federation_id,
-                        amount,
-                        address,
-                    })
-                    .await?;
-
-                print_response(response);
-            }
             Self::ConnectFed {
                 invite_code,
                 #[cfg(feature = "tor")]
@@ -231,11 +164,6 @@ impl GeneralCommands {
                     .await?;
                 print_response(response);
             }
-            Self::Backup { federation_id } => {
-                create_client()
-                    .backup(BackupPayload { federation_id })
-                    .await?;
-            }
             Self::SetConfiguration {
                 password,
                 num_route_hints,
@@ -254,31 +182,6 @@ impl GeneralCommands {
                         per_federation_routing_fees,
                     })
                     .await?;
-            }
-            Self::SpendEcash {
-                federation_id,
-                amount,
-                allow_overpay,
-                timeout,
-                include_invite,
-            } => {
-                let response = create_client()
-                    .spend_ecash(SpendEcashPayload {
-                        federation_id,
-                        amount,
-                        allow_overpay,
-                        timeout,
-                        include_invite,
-                    })
-                    .await?;
-
-                print_response(response);
-            }
-            Self::ReceiveEcash { notes, wait } => {
-                let response = create_client()
-                    .receive_ecash(ReceiveEcashPayload { notes, wait })
-                    .await?;
-                print_response(response);
             }
             Self::Seed => {
                 let response = create_client().get_mnemonic().await?;
