+++ gateway/cli/src/main.rs
@@ -1,15 +1,19 @@
 #![deny(clippy::pedantic, clippy::nursery)]
 
+mod ecash_commands;
 mod general_commands;
 mod lightning_commands;
+mod onchain_commands;
 
 use clap::{CommandFactory, Parser, Subcommand};
+use ecash_commands::EcashCommands;
 use fedimint_core::util::SafeUrl;
 use fedimint_logging::TracingSetup;
 use general_commands::GeneralCommands;
 use lightning_commands::LightningCommands;
 use ln_gateway::rpc::rpc_client::GatewayRpcClient;
 use ln_gateway::rpc::V1_API_ENDPOINT;
+use onchain_commands::OnchainCommands;
 use serde::Serialize;
 
 #[derive(Parser)]
@@ -31,6 +35,10 @@ enum Commands {
     General(GeneralCommands),
     #[command(subcommand)]
     Lightning(LightningCommands),
+    #[command(subcommand)]
+    Ecash(EcashCommands),
+    #[command(subcommand)]
+    Onchain(OnchainCommands),
     Completion {
         shell: clap_complete::Shell,
     },
@@ -47,6 +55,8 @@ async fn main() -> anyhow::Result<()> {
     match cli.command {
         Commands::General(general_command) => general_command.handle(create_client).await?,
         Commands::Lightning(lightning_command) => lightning_command.handle(create_client).await?,
+        Commands::Ecash(ecash_command) => ecash_command.handle(create_client).await?,
+        Commands::Onchain(onchain_command) => onchain_command.handle(create_client).await?,
         Commands::Completion { shell } => {
             clap_complete::generate(
                 shell,
